requirements.txt
Django==3.2.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
django-cors-headers
django-heroku==0.3.1
PyJWT==2.1.0
gunicorn==20.1.0
psycopg2==2.9.1
pandas==1.3.4
numpy==1.21.3
scikit-learn==1.0.1
xgboost==1.5.0
